# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample mealPlans API. You can find out more about the API at https://github.com/PacktPublishing/Building-Distributed-Applications-in-Gin.",
    "title": "Meal Plan API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/mp",
  "paths": {
    "/mp": {
      "get": {
        "description": "Returns list of mealPlans",
        "produces": [
          "application/json"
        ],
        "tags": [
          "mealPlans"
        ],
        "operationId": "list-all"
      },
      "post": {
        "description": "New mealPlan",
        "produces": [
          "application/json"
        ],
        "tags": [
          "mealPlans"
        ],
        "operationId": "create"
      }
    },
    "/mp/{id}": {
      "get": {
        "description": "Get one mealPlan",
        "produces": [
          "application/json"
        ],
        "tags": [
          "mealPlans"
        ],
        "operationId": "describe",
        "parameters": [
          {
            "type": "string",
            "description": "mealPlan ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "put": {
        "description": "Existing mealPlan",
        "produces": [
          "application/json"
        ],
        "tags": [
          "mealPlans"
        ],
        "operationId": "update",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the mealPlan",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      },
      "delete": {
        "description": "Remove an existing mealPlan",
        "produces": [
          "application/json"
        ],
        "tags": [
          "mealPlans"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the mealPlan",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    }
  }
}